move s3 utils into separate file
